Incoporates show controller logic, route, and request tests
With this, all the basic functionality that a user requires should be
implemented Pokemon wise; aside from the seeding of the actual Pokemon.

I doubt we want to allow API based CREATE, UPDATE, or DESTROY access
because the list of existing Pokemon doesn't change frequently and it
shouldn't be managed by the users directly. INDEX and SHOW seem like plenty.